refactor: move utils in hook to csvParserUtils and fix indentation
